Rename import package to imports and move test classes
The rename is because IntelliJ struggles with packages containing the keyword 'import'
