+++ src/main/java/org/purescript/psi/PSExportedItem.kt
@@ -4,7 +4,7 @@ import com.intellij.lang.ASTNode
 import org.purescript.file.PSFile
 import org.purescript.psi.`var`.ExportedModuleReference
 import org.purescript.psi.`var`.ExportedValueReference
-import org.purescript.psi.import.PSImportDeclarationImpl
+import org.purescript.psi.imports.PSImportDeclarationImpl
 
 sealed class PSExportedItem(node: ASTNode) : PSPsiElement(node) {
     val module: PSModule get() = (containingFile as PSFile).module
