+++ src/test/java/org/purescript/psi/imports/ImportedValueReferenceTest.kt
@@ -1,7 +1,6 @@
-package org.purescript.psi.import
+package org.purescript.psi.imports
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
-import junit.framework.TestCase
 import org.purescript.file.PSFile
 
 class ImportedValueReferenceTest : BasePlatformTestCase() {
