+++ src/test/java/org/purescript/psi/imports/PSImportDeclarationImplTest.kt
@@ -1,4 +1,4 @@
-package org.purescript.psi
+package org.purescript.psi.imports
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import junit.framework.TestCase
