+++ src/test/java/org/purescript/psi/imports/PSImportedKindTest.kt
@@ -1,9 +1,8 @@
-package org.purescript.psi
+package org.purescript.psi.imports
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import junit.framework.TestCase
 import org.purescript.file.PSFile
-import org.purescript.psi.import.PSImportedKind
 
 
 class PSImportedKindTest : BasePlatformTestCase() {
