+++ src/test/java/org/purescript/psi/imports/PSImportedValeTest.kt
@@ -1,23 +1,22 @@
-package org.purescript.psi
+package org.purescript.psi.imports
 
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import junit.framework.TestCase
 import org.purescript.file.PSFile
-import org.purescript.psi.import.PSImportedData
 
 
-class PSImportedDataTest : BasePlatformTestCase() {
+class PSImportedValeTest : BasePlatformTestCase() {
 
-    fun `test imported data has correct name`() {
+    fun `test imported value has correct name`() {
         val file = myFixture.configureByText(
             "Foo.purs",
             """
                 module Foo where
-                import Bar (Qux)
+                import Bar (qux)
             """.trimIndent()
         ) as PSFile
         val importDecl = file.module.importDeclarations.single()
-        val importedData = importDecl.importList!!.importedItems.single() as PSImportedData
-        TestCase.assertEquals("Qux", importedData.name)
+        val importedValue = importDecl.importList!!.importedItems.single() as PSImportedValue
+        TestCase.assertEquals("qux", importedValue.name)
     }
 }
